Housekeeping and minor fixes
